+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 	Placeholder for the next version (at the beginning of the line):
 	## __WORK IN PROGRESS__
 -->
-## 5.0.14 (2023-10-17) - Jana
+
+## 5.0.17 (2023-11-29) - Jana
 **BREAKING CHANGES**
 * Support for Node.js 12 and 14 is dropped! Supported are Node.js 16.4.0+ and 18.x
 * Backups created with the new js-controller version cannot be restored on hosts with lower js-controller version!
@@ -21,6 +22,7 @@
 * (foxriver76) Added an admin notification if redis is misconfigured and info how to fix it
 * (foxriver76/buanet) Enable upgrade of js-controller via Admin UI (Linux/Docker only)
 * (foxriver76) automatically reset `info.connection` if adapter instance is stopped
+* (foxriver76) automatically reset instance statistics like `cputime` when instance is stopped
 
 **Optimizations and Fixes**
 * (foxriver76) Speedup "getStates" calls with many IDs often used by visualizations: JSONL 17 times faster, Redis 23 times faster
@@ -48,6 +50,10 @@
 * (foxriver76) fix wrong hostname after backup restore
 * (bluefox) allow CLI vendor update without explicitly specifying vendor file (default file is used)
 * (foxriver76) fix backup restore restoring to old database and using new one afterwards
+* (foxriver76) fix that memory limit of instance was not applied
+* (foxriver76) validate cron jobs before executing them for `schedule` adapters
+* (bluefox) new notification category `securityIssues`
+* (foxriver76) no longer assume all states ending with `.logging` are actual log subscribers
 * (Apollon77, foxriver76, bluefox, AlCalzone) Several fixes and refactorings to prevent potential crash cases reported by Sentry and other sources
 
 **Developer relevant DEPRECATIONS/WARNINGS**